examples/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile dev.in
#
alabaster==0.7.13
    # via sphinx
appdirs==1.4.4
    # via rinohtype
appnope==0.1.3
    # via ipython
asgiref==3.7.2
    # via django
astroid==3.0.0
    # via pylint
asttokens==2.4.0
    # via stack-data
babel==2.13.0
    # via sphinx
backcall==0.2.0
    # via ipython
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
commonmark==0.9.1
    # via recommonmark
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
dill==0.3.7
    # via pylint
django==4.2.6
    # via
    #   -r django_4_2.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-nine
    #   djangorestframework
django-cors-headers==4.2.0
    # via -r django_4_2.in
django-debug-toolbar==4.2.0
    # via -r django_4_2.in
django-debug-toolbar-force==0.2
    # via -r django_4_2.in
django-elasticsearch-debug-toolbar==3.0.2
    # via -r django_4_2.in
django-nine==0.2.7
    # via
    #   django-debug-toolbar-force
    #   django-ormex
django-ormex==0.2.1
    # via -r django_4_2.in
djangorestframework==3.14.0
    # via -r django_4_2.in
docutils==0.20.1
    # via
    #   recommonmark
    #   rinohtype
    #   sphinx
executing==2.0.0
    # via stack-data
gitdb==4.0.10
    # via gitpython
gitpython==3.1.37
    # via matyan
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
ipdb==0.13.13
    # via -r debug.in
ipython==8.16.1
    # via
    #   -r debug.in
    #   ipdb
ipython-genutils==0.2.0
    # via -r debug.in
isort==5.12.0
    # via pylint
jedi==0.19.1
    # via ipython
jinja2==3.1.2
    # via sphinx
markupsafe==2.1.3
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
matyan==0.4.7
    # via -r docs.in
mccabe==0.7.0
    # via pylint
packaging==23.2
    # via
    #   django-nine
    #   rinohtype
    #   sphinx
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==3.11.0
    # via pylint
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.0
    # via -r code_style.in
pygments==2.16.1
    # via
    #   ipython
    #   sphinx
pylint==3.0.1
    # via -r code_style.in
pytz==2023.3.post1
    # via djangorestframework
recommonmark==0.7.1
    # via rinohtype
requests==2.31.0
    # via sphinx
rinoh-typeface-dejavuserif==0.1.3
    # via rinohtype
rinoh-typeface-texgyrecursor==0.1.1
    # via rinohtype
rinoh-typeface-texgyreheros==0.1.1
    # via rinohtype
rinoh-typeface-texgyrepagella==0.1.1
    # via rinohtype
rinohtype==0.5.4
    # via
    #   -r docs.in
    #   rinoh-typeface-dejavuserif
    #   rinoh-typeface-texgyrecursor
    #   rinoh-typeface-texgyreheros
    #   rinoh-typeface-texgyrepagella
six==1.16.0
    # via
    #   asttokens
    #   django-ormex
smmap==5.0.1
    # via gitdb
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.2.6
    # via
    #   -r docs.in
    #   recommonmark
    #   rinohtype
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
tomlkit==0.12.1
    # via pylint
traitlets==5.11.2
    # via
    #   ipython
    #   matplotlib-inline
urllib3==2.0.6
    # via requests
wcwidth==0.2.8
    # via prompt-toolkit
